Compose refactor complete, all scripts working
- Moved cpp/cmp type registration to one place
- Replaced case statement with SymDataType registration for basic
  types
- Altered codegen_api.py to parse this new cpp organization
- ParseData now has map of TypeDesc (datatype, isconst, isref) to
  basic types
- Added ScalarInit logic to allow for simple type casts between scalar
  built in types
- Compose supports function overloading, but requires explciit type
  casts
- Glsl does this the same way, and it makes sense I think. Implicit
  conversions leads to madness.
- Incorpirated cells per type into their registration as wellj
- Parser now sends in DataType enum on basic types instead of string,
  and we lookup the basic registerted type based on DataType enum
- "entity" is now a basic_type, and corresponds to "task_id" on cpp
  side. This is evident in insert_entity system api call
- Appear to have fixed a block packing issue, as shapes.cpp is using
  less blocks than before (2 instead of 4). Logic appears right, but
  don't know how this apparent flaw got fixed, a happy accident of
  refactoring possibly.

+++ src/scripts/cpp/gaen/utils.cpp
@@ -24,7 +24,7 @@
 //   distribution.
 //------------------------------------------------------------------------------
 
-// HASH: 5cc798ac7fd352759900ee248cf9fcd5
+// HASH: 6e41783bbde3699ef82ddde650494f45
 #include "engine/hashes.h"
 #include "engine/Block.h"
 #include "engine/BlockMemory.h"
@@ -245,12 +245,12 @@ public:
         case HASH::init:
         {
             // Params look compatible, message body follows
-            system_api::watch_input_state(HASH::mouse_look, 0, HASH::mouse_look, entity());
+            system_api::watch_input_state(HASH::mouse_look, (u32)0, HASH::mouse_look, entity());
             system_api::watch_mouse(HASH::mouse_move, HASH::mouse_wheel, entity());
-            system_api::watch_input_state(HASH::forward, 0, HASH::forward, entity());
-            system_api::watch_input_state(HASH::back, 0, HASH::back, entity());
-            system_api::watch_input_state(HASH::left, 0, HASH::left, entity());
-            system_api::watch_input_state(HASH::right, 0, HASH::right, entity());
+            system_api::watch_input_state(HASH::forward, (u32)0, HASH::forward, entity());
+            system_api::watch_input_state(HASH::back, (u32)0, HASH::back, entity());
+            system_api::watch_input_state(HASH::left, (u32)0, HASH::left, entity());
+            system_api::watch_input_state(HASH::right, (u32)0, HASH::right, entity());
             return MessageResult::Consumed;
         }
         case HASH::mouse_look:
@@ -487,12 +487,12 @@ public:
         case HASH::init:
         {
             // Params look compatible, message body follows
-            system_api::watch_input_state(HASH::mouse_look, 0, HASH::mouse_look, entity());
+            system_api::watch_input_state(HASH::mouse_look, (u32)0, HASH::mouse_look, entity());
             system_api::watch_mouse(HASH::mouse_move, HASH::mouse_wheel, entity());
-            system_api::watch_input_state(HASH::forward, 0, HASH::forward, entity());
-            system_api::watch_input_state(HASH::back, 0, HASH::back, entity());
-            system_api::watch_input_state(HASH::left, 0, HASH::left, entity());
-            system_api::watch_input_state(HASH::right, 0, HASH::right, entity());
+            system_api::watch_input_state(HASH::forward, (u32)0, HASH::forward, entity());
+            system_api::watch_input_state(HASH::back, (u32)0, HASH::back, entity());
+            system_api::watch_input_state(HASH::left, (u32)0, HASH::left, entity());
+            system_api::watch_input_state(HASH::right, (u32)0, HASH::right, entity());
             return MessageResult::Consumed;
         }
         case HASH::forward:
